+++ packages/browser-wallet/src/popup/pages/Account/TransactionLog/i18n/en.ts
@@ -7,6 +7,7 @@ const t = {
     events: 'Events',
     rejectReason: 'Reject reason',
     noTransactions: 'This account has no transactions yet.',
+    error: 'Unable to retrieve transaction history',
 };
 
 export default t;